backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Image Processing
Pillow==10.2.0
rawpy==0.19.0
imageio==2.33.1
exifread==3.0.0

# AI/ML
torch==2.1.2
torchvision==0.16.2
transformers==4.37.0

# HTTP Client (for LLM API calls)
httpx==0.26.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1

# Geocoding / Mapping
geopy==2.4.1
folium==0.15.1

# Excel Export
openpyxl==3.1.2
